start the process of making matches out of requests, add offline tournament to ELO object, still a lot to do
