+++ lm_tournament_eval/api/elo.py
@@ -10,13 +10,8 @@ class ELO:
         self.score_0 = 1200
         self.score_1 = 1200
         self.k = 16
-        self.match_result_list = [MatchResult(model0_old_elo=self.score_0,
-                                              model1_old_elo=self.score_1,
-                                              model0_new_elo=self.score_0,
-                                              model1_new_elo=self.score_1)]
 
-
-    def elo_update(self, results0 : dict, results1 : dict, task_names : List):
+    def online_elo_update(self, results0 : dict, results1 : dict, task_names : List):
         # self.match_result_list.model0_old_elo = self.score_0
         # self.match_result_list.model1_old_elo = self.score_1
         print(f"score_0, 1 {self.score_0}, {self.score_1}")
@@ -76,3 +71,36 @@ class ELO:
         print("----------------------------")
         # self.match_result_list[n].model0_new_elo = self.score_0
         # self.match_result_list[n].model1_new_elo = self.score_1
+
+    def offline_elo_update(results0, results1, task_names, task_indices):
+        # run match
+        # sample indices
+        # calculate the wins, losses, and draws
+        as_1 = []
+        as_2 = []
+        for i in task_indices:
+            # draw
+            if results0[i]['acc'] == results1[i]['acc']:
+                as_1.append(0)
+                as_2.append(0)
+            # model 1 won 
+            elif results0[i]['acc'] > results1[i]['acc']:
+                as_1.append(1)
+            # model 2 won
+            elif results0[i]['acc'] < results1[i]['acc']:
+                as_2.append(1)
+        expected_score_0 = 1/(1+10**((self.score_0-self.score_1)/400))
+        expected_score_1 = 1/(1+10**((self.score_1-self.score_0)/400))
+        
+        # update Elo
+        if sum(as_1) > sum(as_2):
+            self.score_0 = self.score_0 + self.k*(1 - expected_score_0)
+            self.score_1 = self.score_1 + self.k*(0 - expected_score_1)
+        elif sum(as_2) > sum(as_1):
+            self.score_0 = self.score_0 + self.k*(0 - expected_score_0)
+            self.score_1 = self.score_1 + self.k*(1 - expected_score_1)
+        elif sum(as_1) == sum(as_2):
+            self.score_0 = self.score_0 + self.k*(0.5 - expected_score_0)
+            self.score_1 = self.score_1 + self.k*(0.5 - expected_score_1)           
+        print(f"score_0, 1 {self.score_0}, {self.score_1}")
+        print("----------------------------")
